requirements.txt
azure-search-documents
azure-storage-blob
python-dotenv
pypdf
streamlit
